chore(grunt): Add grunt to devDependencies
This change allows us to install grunt-cli (`npm install -g grunt-cli`).
Grunt-cli will run a local version of gruntJS if there is one, falling back to the global one otherwise.
This allows us to run different versions of grunt on a project by project basis - particularly important as Angular.js looks like it is moving to Grunt 0.4.0 at some point.

So if you have installed grunt-cli you now need to do an npm install inside the client folder to ensure that the correct version of grunt is loaded.