test(admin-projects): Remove artificial array mocks.
Mocking up the array object in $scope.project.teamMembers was too artificial and would make the tests too brittle.
I now just set up the array manually.  I am allowed to do this because $scope.project.teamMembers is a public API of this controller - in the sense that it is used in the template as-is.