fix(login): Move form element to enable submit on enter.
The primary default button for the login form was outside the form element.
By moving the form up a level in the DOM the button becomes part of the form and so will submit the form when the user hits enter in one of the input boxes.
Since the form had more than two input boxes, the ng-submit on the form was never being triggered.  Removed now to prevent double submission.