Fixes problem with scope where the selTeamMember value in scope for our HTML elements was not the same as the value on $scope in the ctrl.
Adding users was not working because $scope.selTeamMember was always
undefined in the controller. However, bindings on the "selTeamMember"
model in the HTML worked correctly. So there must be an implicit sibling
scope involved for some reason. Changing the model to be an object
instead removes ambiguity for Angular and the interaction now works as
expected.